fix(src/reactpictureannotation.tsx): change of selectedId didn't changne selected transformer box
When change of selectedId was triggered programmatically by the application logic: only the
selection of the box and display of comment box took place, Transformer box was always stuck on
previously selected annotation. This did not happen when selection was changed by selecting new box
with the mouse, because Transformer was set in that scenario in DefaultAnnotationState.onMouseDown.
This was misleading the user. I added Next button in the storybook to reproduce the faulty behavior.